requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
transformers>=5.3.0
torch>=2.1.0
Pillow>=10.0.0
requests>=2.31.0
python-multipart>=0.0.9
pydantic>=2.0.0
pymupdf>=1.24.0
accelerate>=0.30.0